Add Login Create

+++ WSTowersOffice.Api/Controllers/AdminsController.cs
@@ -8,6 +8,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using WSTowersOffice.Api.Models;
+using ShowProducts.API.Controllers;
 
 namespace WSTowersOffice.Api.Controllers
 {
@@ -18,12 +19,22 @@ namespace WSTowersOffice.Api.Controllers
         // GET: Admins
         public async Task<ActionResult> Manager()
         {
+            var loginResult = await LoginController.ValidLoginAsync();
+            if (!loginResult.IsValid)
+            {
+                return RedirectToActionPermanent("Authentication", "Logins");
+            }
             return View(await db.Login.ToListAsync());
         }
 
         // GET: Admins/Details/5
         public async Task<ActionResult> Details(int? id)
         {
+            var loginResult = await LoginController.ValidLoginAsync();
+            if (!loginResult.IsValid)
+            {
+                return RedirectToActionPermanent("Authentication", "Logins");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -37,8 +48,13 @@ namespace WSTowersOffice.Api.Controllers
         }
 
         // GET: Admins/Create
-        public ActionResult Create()
+        public async Task<ActionResult> Create()
         {
+            var loginResult = await LoginController.ValidLoginAsync();
+            if (!loginResult.IsValid)
+            {
+                return RedirectToActionPermanent("Authentication", "Logins");
+            }
             return View();
         }
 
@@ -47,11 +63,48 @@ namespace WSTowersOffice.Api.Controllers
         // Para obter mais detalhes, confira https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create([Bind(Include = "ID,UserName,Password,Email,isValidEmail,CreateDate,ValidKey")] Login login)
+        public async Task<ActionResult> Create([Bind(Include = "ID,UserName,Password,Email,isValidEmail,CreateDate,ValidKey")] LoginModel login)
         {
+            var loginResult = await LoginController.ValidLoginAsync();
+            if (!loginResult.IsValid)
+            {
+                return RedirectToActionPermanent("Authentication", "Logins");
+            }
+
+            if ((await db.Login.FirstOrDefaultAsync(fs => fs.Email == login.Email)) != null)
+            {
+                ModelState.AddModelError("Email", "This 'e-mail' already ben using.");
+            }
+
+            if ((await db.Login.FirstOrDefaultAsync(fs => fs.UserName == login.UserName)) != null)
+            {
+                ModelState.AddModelError("Username", "This 'Username' already ben using.");
+            }
+
+            Login loginModel = login.GetLogin();
+
+            if (login.Password != login.ConfirmPassword)
+            {
+                ModelState.AddModelError("ConfirmPassword", "Password confirmation of being equal to password");
+            }
+            loginModel.Password = LoginController.CryptographyString(login.Password);
+            loginModel.CreateDate = DateTime.UtcNow;
+
+            Guid userkey = Guid.NewGuid();
+            bool exist = false;
+            do
+            {
+                if ((await db.Login.FirstOrDefaultAsync(fs => fs.ValidKey == userkey.ToString())) == null)
+                {
+                    exist = true;
+                    userkey = Guid.NewGuid();
+                }
+            } while (exist);
+
+            loginModel.ValidKey = userkey.ToString();
             if (ModelState.IsValid)
             {
-                db.Login.Add(login);
+                db.Login.Add(loginModel);
                 await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
@@ -62,6 +115,11 @@ namespace WSTowersOffice.Api.Controllers
         // GET: Admins/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
+            var loginResult = await LoginController.ValidLoginAsync();
+            if (!loginResult.IsValid)
+            {
+                return RedirectToActionPermanent("Authentication", "Logins");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -81,6 +139,11 @@ namespace WSTowersOffice.Api.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit([Bind(Include = "ID,UserName,Password,Email,isValidEmail,CreateDate,ValidKey")] Login login)
         {
+            var loginResult = await LoginController.ValidLoginAsync();
+            if (!loginResult.IsValid)
+            {
+                return RedirectToActionPermanent("Authentication", "Logins");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(login).State = EntityState.Modified;
@@ -93,6 +156,11 @@ namespace WSTowersOffice.Api.Controllers
         // GET: Admins/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
+            var loginResult = await LoginController.ValidLoginAsync();
+            if (!loginResult.IsValid)
+            {
+                return RedirectToActionPermanent("Authentication", "Logins");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -110,6 +178,11 @@ namespace WSTowersOffice.Api.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
+            var loginResult = await LoginController.ValidLoginAsync();
+            if (!loginResult.IsValid)
+            {
+                return RedirectToActionPermanent("Authentication", "Logins");
+            }
             Login login = await db.Login.FindAsync(id);
             db.Login.Remove(login);
             await db.SaveChangesAsync();
